database edition 0.3

+++ db/migrate/20230324124202_drop_events.rb
@@ -1,5 +1,5 @@
 class DropEvents < ActiveRecord::Migration[6.1]
   def change
-    drop_table :events
+    drop_table :events if ActiveRecord::Base.connection.table_exists?(:events)
   end
 end
