+++ db/migrate/20230324124510_drop_fighters.rb
@@ -1,5 +1,5 @@
 class DropFighters < ActiveRecord::Migration[6.1]
   def change
-    drop_table :fighters
+    drop_table :fighters if ActiveRecord::Base.connection.table_exists?(:fighters)
   end
 end
