+++ db/migrate/20230404012520_drop_fighter_photos.rb
@@ -1,5 +1,5 @@
 class DropFighterPhotos < ActiveRecord::Migration[6.1]
   def change
-    drop_table :fighter_photos
+    drop_table :fighter_photos if ActiveRecord::Base.connection.table_exists?(:fighter_photos)
   end
 end
